Created jpa database

+++ app/src/main/java/pl/futurecollars/invoicing/controllers/invoice/InvoiceApi.java
@@ -23,17 +23,17 @@ public interface InvoiceApi {
 
     @ApiOperation(value = "Method used to add invoice")
     @PostMapping
-    ResponseEntity<Integer> add(@RequestBody Invoice invoice);
+    ResponseEntity<Long> add(@RequestBody Invoice invoice);
 
     @ApiOperation(value = "Method used to get invoice by Id")
     @GetMapping("/{id}")
-    ResponseEntity<Invoice> getById(@PathVariable int id);
+    ResponseEntity<Invoice> getById(@PathVariable long id);
 
     @ApiOperation(value = "Method used to update invoice by Id ")
     @PutMapping("/{id}")
-    ResponseEntity<Invoice> updateById(@PathVariable int id, @RequestBody Invoice invoice);
+    ResponseEntity<Invoice> updateById(@PathVariable long id, @RequestBody Invoice invoice);
 
     @ApiOperation(value = "Method used to delete invoice by Id")
     @DeleteMapping("/{id}")
-    ResponseEntity<Invoice> deleteById(@PathVariable int id);
+    ResponseEntity<Invoice> deleteById(@PathVariable long id);
 }
