+++ app/src/main/java/pl/futurecollars/invoicing/db/memory/InMemoryDatabase.java
@@ -12,17 +12,17 @@ import pl.futurecollars.invoicing.model.Invoice;
 public class InMemoryDatabase implements Database {
 
     private Map<Integer, Invoice> memoryDatabase = new HashMap<>();
-    private int count = 1;
+    private long count = 1;
 
     @Override
-    public int save(Invoice invoice) {
-        this.memoryDatabase.put(count, invoice);
+    public long save(Invoice invoice) {
+        this.memoryDatabase.put((int) count, invoice);
         invoice.setId(count);
         return count++;
     }
 
     @Override
-    public Optional<Invoice> getByID(int id) {
+    public Optional<Invoice> getByID(long id) {
         printIllegalArgumentException(id);
         return memoryDatabase.values()
             .stream()
@@ -38,20 +38,20 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         printIllegalArgumentException(id);
-        memoryDatabase.replace(id, memoryDatabase.get(id), updatedInvoice);
+        memoryDatabase.replace((int) id, memoryDatabase.get(id), updatedInvoice);
         updatedInvoice.setId(id);
         return Optional.of(updatedInvoice);
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         printIllegalArgumentException(id);
         return Optional.ofNullable(memoryDatabase.remove(id));
     }
 
-    private void printIllegalArgumentException(int id) {
+    private void printIllegalArgumentException(long id) {
         if (id <= 0) {
             throw new IllegalArgumentException("Error: id cannot be negative");
         }
